requirements.txt
click>=7.1.2
consolekit>=0.5.0
deprecation-alias>=0.1.1
domdf-python-tools>=2.0.0
dulwich!=0.20.7,!=0.20.8,!=0.20.9,>=0.20.5
github3-py>=1.3.0
github3-utils>=0.1.1
repo-helper>=2020.12.18
southwark>=0.4.0
typing-extensions>=3.7.4.3
